add jquery

+++ server/scripts/celery.py
@@ -56,5 +56,6 @@ except KeyboardInterrupt:
     proc.send_signal(signal.SIGTERM)
     time = datetime.now()
     # rename the log file
-    copyfile(logfile, f"logs/log{time}.log")
+    if os.path.exists(logfile):
+        copyfile(logfile, f"logs/log{time}.log")
     exit(0)
